refactor(agents): split OpenCode ACP protocol parsing
